Added Arrows
Added arrows by the descriptions so that users can move that way. Cycle timer is now 15 seconds, and the counter resets if a user clicks and arrow, or clicks directly on one of the steps in the graphic. Not sure if I like how it fades in and out, and may change it to "slide" in the appropriate direction