Having an issue...
Pushing this so you can also take a look. I renamed the steps in the splurge that were leftover names from when I forked it: mergeA-description -> splurgeA-description, for example.

Additionally, I changed the logic in status.js to look for the correct name to map the steps to the descriptions (see code starting at line 68). I haven't upgraded the logic to go beyond step L yet.

I can't figure out why it's not working...thought maybe you might have an idea. Trying to debug with chatgpt