Fixed description logic!
Changed the description logic so that it will work with each page. It is better at determining and matching the step ID to the step description, and will dynamically cycle to the next letter in the alphabet irrespective of how many steps there are.

Experimented with adding keyboard arrow functionality, but ran into a few problems. Will likely add later